+++ R/clinicaltrials_gov_version.R
@@ -10,10 +10,6 @@
 #'     recent, etc. If no version number is specified, the first
 #'     version will be downloaded.
 #'
-#' @param polite If TRUE, this function uses the `polite` package to
-#'     download data, if FALSE, this function uses the `rvest`
-#'     package, default TRUE.
-#'
 #' @return A list containing the overall status, enrolment, start
 #'     date, start date precision (month or day) primary completion
 #'     date, primary completion date precision (month or day), primary
@@ -35,8 +31,7 @@
 #'
 clinicaltrials_gov_version <- function(
                                        nctid,
-                                       versionno=1,
-                                       polite=TRUE
+                                       versionno=1
                                        ) {
 
     out <- tryCatch({
@@ -69,12 +64,8 @@ clinicaltrials_gov_version <- function(
             versionno
         )
 
-        if (polite) {
-            version <- polite_read_html(url)
-        } else {
-            version <- rvest::read_html(url)
-        }
-
+        version <- rvest::read_html(url)
+        
         ## Back up locale info
         lct <- Sys.getlocale("LC_TIME")
         ## Set locale so that months are parsed correctly on