+++ tests/testthat/test-clinicaltrials_gov_version.R
@@ -1,5 +1,5 @@
 test_that("ClinicalTrials.gov version downloads correctly", {
-    result <- clinicaltrials_gov_version("NCT00942747", 1, polite=FALSE)
+    result <- clinicaltrials_gov_version("NCT00942747", 1)
     expect_equal(
         length(result) == 19 || result == "Error",
         TRUE