+++ tests/testthat/test-clinicaltrials_gov_version_references.R
@@ -1,5 +1,5 @@
 test_that("ClinicalTrials.gov references downloads correctly", {
-    version <- clinicaltrials_gov_version("NCT04315480", 3, polite=FALSE)
+    version <- clinicaltrials_gov_version("NCT04315480", 3)
     if (version[1] == "Error") {
         expect_equal(
             version[1],