+++ tests/testthat/test-clinicaltrials_gov_version_startdate.R
@@ -1,5 +1,5 @@
 test_that("ClinicalTrials.gov start date downloads correctly", {
-    version <- clinicaltrials_gov_version("NCT02110043", 1, polite=FALSE)
+    version <- clinicaltrials_gov_version("NCT02110043", 1)
     if (version[1] == "Error") {
         expect_equal(
             version[1],